+++ test/debug-mode.lisp
@@ -1,4 +1,0 @@
-
-(in-package weblocks-test)
-
-; All tests were deprecated, waiting for new