+++ ChangeLog.rst
@@ -2,6 +2,16 @@
  ChangeLog
 ===========
 
+Unreleased
+==========
+
+Fixes
+-----
+
+* Fixed issue introduced in ``0.33.2`` release - system
+  ``lack-middleware-session`` was loaded successfully, but SBCL was
+  unable to find ``lack-middleware-session`` package.
+
 0.33.2 (2018-12-06)
 ===================